fix: fix identity hash from Digest calls
The identity hash needs a size of the actual hash associated with the
digest, hence ut needs custom Digest trait implementations.
